refactor src/gpu -> src/gfx
